pyproject.toml
[tool.poetry]
name = "gshare"
version = "0.1.0"
description = "A Backend for extracting subtitles from uploaded videos"
authors = ["Gagandeep Singh Bhambrah <grimmgagan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pre-commit = "^3.8.0"
django-stubs = "^5.0.4"
djangorestframework-stubs = "^3.15.0"
celery-stubs = "^0.1.3"
pyright = "^1.1.379"
pytest-env = "^1.1.4"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env = [
    "DJANGO_SETTINGS_MODULE=config.settings",
]
